사소한 리팩토링 2
