--- src/lib/components/molecules/Categories/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Categories.svelte";
-export * from "./service";
