--- src/lib/components/molecules/Categories/service.ts
@@ -1,7 +1,0 @@
-import type { DataKey } from "$lib/modules/filesystem";
-
-export interface SelectedCategory {
-  id: number;
-  dataKey?: DataKey;
-  name: string;
-}
