+++ src/lib/components/molecules/index.ts
@@ -1,7 +1,7 @@
 export * from "./ActionModal.svelte";
 export { default as ActionModal } from "./ActionModal.svelte";
-export * from "./Categories";
-export { default as Categories } from "./Categories";
+export * from "./Categories.svelte";
+export { default as Categories } from "./Categories.svelte";
 export { default as IconEntryButton } from "./IconEntryButton.svelte";
 export * from "./labels";
 export { default as SubCategories } from "./SubCategories.svelte";
